Removed bug with bodies by only updating when the body is not destroyed

+++ core/src/objects/Bullet.java
@@ -33,16 +33,21 @@ public class Bullet extends MovingRectangle{
 
     @Override
     public void update(){
-        if(lives <= 0 || lastX == x || outOfSight) {
+        if((lives <= 0 || lastX == x || outOfSight) && !destroyed) {
             gameScreen.removeMovingRectangle(this);
             gameScreen.getWorld().destroyBody(this.body);
+            body.setUserData(null);
+            body = null;
+            destroyed = true;
             return;
         }
 
-        lastX = x;
-        super.update();
-        body.setLinearVelocity(directionX*speedLevel*speed, directionY*speedLevel*speed);
-        outOfSight = bulletOutOfSight();
+        else if(!destroyed) {
+            lastX = x;
+            super.update();
+            body.setLinearVelocity(directionX*speedLevel*speed, directionY*speedLevel*speed);
+            outOfSight = bulletOutOfSight();
+        }
     }
 
     public void setRemove(boolean remove) {
